Weapons Fully Functional
Weapons added in fully and tested on 3 different devices between @sagefantasma, @snakeswiss and myself.

Strings of every area in the game added in to a StringManager class similar to how I did it in the MGS3 trainer.

Just need to confirm offsets for all the items, life and where the area string is and then implement ClickOnce and we'll be set for V1.0 of the MGS1 trainer.